// src/enums/calendar.ts
export enum CalendarActionTypes {
    FETCH_CALENDAR_BOOKINGS_REQUEST = '@@calendar/FETCH_CALENDAR_BOOKINGS_REQUEST',
    FETCH_CALENDAR_BOOKINGS_SUCCESS = '@@calendar/FETCH_CALENDAR_BOOKINGS_SUCCESS',
    FETCH_CALENDAR_BOOKINGS_FAILURE = '@@calendar/FETCH_CALENDAR_BOOKINGS_FAILURE',
    CHANGE_SELECTED_DATE = '@@calendar/CHANGE_SELECTED_DATE',
    ADD_FILTER = '@@calendar/ADD_FILTER',
    REMOVE_FILTER = '@@calendar/REMOVE_FILTER',
    CHANGE_MONTH = '@@calendar/CHANGE_MONTH',
    SELECT_BOOKING = '@@calendar/SELECT_BOOKING',
}

export enum FilterTypes {
    SHOW_PENDING = 'SHOW_PENDING',
    SHOW_COMPLETED = 'SHOW_COMPLETED',
    SHOW_MOVING_REQUESTS = 'SHOW_MOVING_REQUESTS',
    SHOW_IN_PROCESS = 'SHOW_IN_PROCCESS',
};